// src/apis/generated/model/optionalProcedureRes.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * 동행 API
 * 동행 서비스 API 입니다.
 * OpenAPI spec version: v1
 */

export interface OptionalProcedureRes {
  procedureId?: number;
  procedureName?: string;
  categoryId?: number;
  categoryName?: string;
  remainingDays?: number;
  priority?: string;
}